+++ src/main/java/com/ojo/cyrus/models/responses/PaymentEventListItem.java
@@ -5,7 +5,7 @@ import com.ojo.cyrus.enums.NombaPaymentEventType;
 import com.ojo.cyrus.enums.ReconciliationFailureReason;
 import io.swagger.v3.oas.annotations.media.Schema;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -30,7 +30,7 @@ public record PaymentEventListItem(
         @Schema(description = "Human-readable details about the processing outcome")
         String statusDetails,
         @Schema(description = "Amount in kobo (if available from the raw payload)")
-        BigInteger amount,
+        BigDecimal amount,
         @Schema(description = "Virtual account number the payment was sent to (if identifiable)")
         String accountNumber,
         @Schema(description = "Customer reference this event was attributed to (null if still orphaned)")
