+++ src/main/java/com/ojo/cyrus/models/responses/StatementRowResponse.java
@@ -1,6 +1,6 @@
 package com.ojo.cyrus.models.responses;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.Instant;
 
 /** One line of a customer's statement — a single inbound transaction. */
@@ -9,5 +9,5 @@ public record StatementRowResponse(
         String payer,
         String ref,
         String matchStatus,
-        BigInteger amountKobo
+        BigDecimal amountKobo
 ) {}
